tetragon: Move config package to pkg/config
Let's make config package global and move stuff in it in following
changes.
