+++ pkg/policyfilter/namespace.go
@@ -11,8 +11,8 @@ import (
 
 	"github.com/cilium/ebpf"
 	"github.com/cilium/tetragon/pkg/bpf"
+	"github.com/cilium/tetragon/pkg/config"
 	"github.com/cilium/tetragon/pkg/option"
-	"github.com/cilium/tetragon/pkg/sensors/exec/config"
 	lru "github.com/hashicorp/golang-lru/v2"
 )
